working on assignment-37

+++ lib/random_data.rb
@@ -28,6 +28,6 @@ module RandomData
 
 
    def rand_price
-     rand(10..50).to_i
+     rand(10..50).to_f
    end
 end
